Fixed 'TypeError: Got unsupported ScalarType BFloat16' in PPO trainer when logging and using BF16 during training

+++ trl/trainer/ppo_trainer.py
@@ -1355,9 +1355,9 @@ class PPOTrainer(BaseTrainer):
                 if isinstance(v, torch.Tensor) and v.dtype == torch.bfloat16:
                     logs[k] = v.float()
 
-            logs["env/reward_mean"] = torch.mean(rewards).cpu().numpy().item()
-            logs["env/reward_std"] = torch.std(rewards).cpu().numpy().item()
-            logs["env/reward_dist"] = rewards.cpu().numpy()
+            logs["env/reward_mean"] = torch.mean(rewards.type(torch.float)).cpu().numpy().item()
+            logs["env/reward_std"] = torch.std(rewards.type(torch.float)).cpu().numpy().item()
+            logs["env/reward_dist"] = rewards.type(torch.float).cpu().numpy()
 
             if self.config.log_with == "tensorboard":
                 # update the current step
